Add setup.py, rename source folder
