+++ brick_sequencer/Accuracy.py
@@ -1,6 +1,6 @@
-from src.BaseGuesser import *
-import src.BaseCaller
-from src.TimeCalculator import *
+from brick_sequencer.BaseGuesser import *
+import brick_sequencer.BaseCaller
+from brick_sequencer.TimeCalculator import *
 
 from os import walk
 
@@ -49,7 +49,7 @@ def getAccuracy(dir, files):
 
             #print(time)
 
-            dfRes = BaseGuesser.sequence(src.BaseCaller.dropNonSequence(df), time)
+            dfRes = BaseGuesser.sequence(brick_sequencer.BaseCaller.dropNonSequence(df), time)
             accuracy = calculateAcurracyRun(sequence, dfRes)
             print(accuracy)
             print(dfRes)
