+++ brick_sequencer/colours.py
